refactor: move spork related code from llmq/utils to llmq/options
